Disable sixel mode on reset, screen clear.
Previously, if someone entered sixel mode and accidentally didn't exit
it, the screen would just appear to be frozen. Nothing could be done
other than switch to another terminal to kill yaft.

Xterm doesn't have this problem because it turns off sixel mode when
the screen clears (user hits ^L) and during a reset. This is easily
accomplished by changing the terminfo for yaft to send '\E\\' (end
sixel mode).