titeInhibit: Prevent alternate screen switching.
Short: Set smcup and rmcup to something harmless so that curses
doesn't think they're accidentally missing and try to fake an
alternate screen.

Long: Alternate screen switching is for terminals that need to be in a
special mode before they can use cursor motions, like the Tektronix
4025. On modern terminals, it just appears to reset the screen when a
program is entered or exited. This can be annoying: ever looked
something up with 'w3m' or 'man' and then wanted to type it into the
terminal? If your terminal has "alternate screen switching" enabled
the text will be cleared when exiting the program and you cannot read
it.

When using X or Wayland, a person can work around this annoyance by
having multiple terminals open at once, one to read, one to type in.
But with yaft, we can't do that.

Yaft is correct in not defining smcup or rmcup. Theoretically it
shouldn't suffer from alternate screen switching problems.
Unfortunately, the ncurses library appearsto fake smcup and rmcup
calls if they are not defined for the terminal!

So, as a workaround, this patch defines smcup and rmcup to be '\r'
(ASCII carriage return). This is a harmless character to send as it
simply puts the cursor at the beginning of the line.

By the way, I refer to this as "titeInhibit", because that is what Xterm
calls the feature to disable "alternate screen switching". The termcap
codes for initializing and exiting it are "ti" and "te".